adds getMultiFile and prevents UI blocking when picking large remote files
