+++ android/src/main/java/com/mr/flutter/plugin/filepicker/FilePickerPlugin.java
@@ -52,52 +52,58 @@ public class FilePickerPlugin implements MethodCallHandler {
     instance = registrar;
     instance.addActivityResultListener(new PluginRegistry.ActivityResultListener() {
       @Override
-      public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
+      public boolean onActivityResult(int requestCode, int resultCode, final Intent data) {
 
         if (requestCode == REQUEST_CODE && resultCode == Activity.RESULT_OK) {
 
-          if (data != null) {
-            if(data.getClipData() != null) {
-              int count = data.getClipData().getItemCount();
-              int currentItem = 0;
-              ArrayList<String> paths = new ArrayList<>();
-              while(currentItem < count) {
-                final Uri currentUri = data.getClipData().getItemAt(currentItem).getUri();
-                String path = FileUtils.getPath(currentUri, instance.context());
-                if(path == null) {
-                  path =  FileUtils.getUriFromRemote(instance.activeContext(), currentUri, result);
+            new Thread(new Runnable() {
+                @Override
+                public void run() {
+                    if (data != null) {
+                        if(data.getClipData() != null) {
+                            int count = data.getClipData().getItemCount();
+                            int currentItem = 0;
+                            ArrayList<String> paths = new ArrayList<>();
+                            while(currentItem < count) {
+                                final Uri currentUri = data.getClipData().getItemAt(currentItem).getUri();
+                                String path = FileUtils.getPath(currentUri, instance.context());
+                                if(path == null) {
+                                    path =  FileUtils.getUriFromRemote(instance.activeContext(), currentUri, result);
+                                }
+                                paths.add(path);
+                                Log.i(TAG, "[MultiFilePick] File #" + currentItem + " - URI: " +currentUri.getPath());
+                                currentItem++;
+                            }
+                            if(paths.size() > 1){
+                                runOnUiThread(result, paths, true);
+                            } else {
+                                runOnUiThread(result, paths.get(0), true);
+                            }
+                        } else if (data.getData() != null) {
+                            Uri uri = data.getData();
+                            Log.i(TAG, "[SingleFilePick] File URI:" + uri.toString());
+                            String fullPath = FileUtils.getPath(uri, instance.context());
+
+                            if(fullPath == null) {
+                                fullPath =  FileUtils.getUriFromRemote(instance.activeContext(), uri, result);
+                            }
+
+                            if(fullPath != null) {
+                                Log.i(TAG, "Absolute file path:" + fullPath);
+                                runOnUiThread(result, fullPath, true);
+                            } else {
+                                runOnUiThread(result, "Failed to retrieve path.", false);
+                            }
+                        } else {
+                            runOnUiThread(result, "Unknown activity error, please fill an issue.", false);
+                        }
+                    } else {
+                        runOnUiThread(result, "Unknown activity error, please fill an issue.", false);
+                    }
                 }
-                paths.add(path);
-                Log.i(TAG, "[MultiFilePick] File #" + currentItem + " - URI: " +currentUri.getPath());
-                currentItem++;
-              }
-              if(paths.size() > 1){
-                result.success(paths);
-              } else {
-                result.success(paths.get(0));
-              }
-            } else if (data.getData() != null) {
-              Uri uri = data.getData();
-              Log.i(TAG, "[SingleFilePick] File URI:" + uri.toString());
-              String fullPath = FileUtils.getPath(uri, instance.context());
-  
-              if(fullPath == null) {
-               fullPath =  FileUtils.getUriFromRemote(instance.activeContext(), uri, result);
-              }
-  
-              if(fullPath != null) {
-                Log.i(TAG, "Absolute file path:" + fullPath);
-                result.success(fullPath);
-              } else {
-                result.error(TAG, "Failed to retrieve path." ,null);
-              }
-            } else {
-              result.error(TAG, "Unknown activity error, please fill an issue." ,null);
-            }
-          } else {
-            result.error(TAG, "Unknown activity error, please fill an issue." ,null);
-          }
-          return true;
+            }).start();
+            return true;
+
         } else if(requestCode == REQUEST_CODE && resultCode == Activity.RESULT_CANCELED) {
             result.success(null);
             return true;
@@ -120,6 +126,21 @@ public class FilePickerPlugin implements MethodCallHandler {
       }
     });
   }
+
+  private static void runOnUiThread(final Result result, final Object o, final boolean success) {
+        instance.activity().runOnUiThread(new Runnable() {
+            @Override
+            public void run() {
+                if(success) {
+                    result.success(o);
+                } else if(o != null) {
+                    result.error(TAG,(String)o, null);
+                } else {
+                    result.notImplemented();
+                }
+            }
+        });
+    }
 
   @Override
   public void onMethodCall(MethodCall call, Result result) {
@@ -183,12 +204,8 @@ public class FilePickerPlugin implements MethodCallHandler {
     Intent intent;
 
     if (checkPermission()) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.KITKAT) {
-            intent = new Intent(Intent.ACTION_PICK);
-        } else {
-            intent = new Intent(Intent.ACTION_GET_CONTENT);
-        }
 
+        intent = new Intent(Intent.ACTION_GET_CONTENT);
         Uri uri = Uri.parse(FileUtils.getExternalPath(instance.activeContext()) + File.separator);
         intent.setDataAndType(uri, type);
         intent.setType(type);
