+++ example/lib/main_desktop.dart
@@ -1,0 +1,8 @@
+import 'package:file_picker_example/src/file_picker_demo.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/widgets.dart';
+
+void main() {
+  debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
+  runApp(new FilePickerDemo());
+}
