+++ pubspec.yaml
@@ -4,7 +4,6 @@ author: Miguel Ruivo <miguel@miguelruivo.com>
 homepage: https://github.com/miguelpruivo/plugins_flutter_file_picker
 version: 1.4.0
 
-
 dependencies:
   flutter:
     sdk: flutter
@@ -12,10 +11,6 @@ dependencies:
 environment:
   sdk: ">=2.0.0 <3.0.0"
 
-# For information on the generic Dart part of this file, see the
-# following page: https://www.dartlang.org/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
   plugin:
     androidPackage: com.mr.flutter.plugin.filepicker
